deleted some unnecessary code
