+++ pages/index.js
@@ -1,9 +1,23 @@
-import ListPage from "./ListPage";
+import CampingCard from "../components/CampingCard";
+import { campingTrips } from "../data/campingTrips";
+import { Container, Card, Header, Footer } from "../styles/styles";
 
-export default function HomePage() {
+export default function ListPage() {
   return (
-    <div>
-      <ListPage />
-    </div>
+    <>
+      <Header>
+        <h1>Organize your Trips</h1>
+      </Header>
+      <Container>
+        {campingTrips.map((trip) => (
+          <Card key={trip.id}>
+            <CampingCard trip={trip} />
+          </Card>
+        ))}
+      </Container>
+      <Footer>
+        <p>FOOTER</p>
+      </Footer>
+    </>
   );
 }
